op-acceptance-tests: dedupe ZK honest-challenger tests across sources
The op-node and supernode ZK challenger tests were identical except for
which super-root source the honest challenger reads from. Collapse each
scenario (valid, invalid, unsafe) into a single test with op-node and
supernode subtests that share one scenario body, so the migration is
proven identical for both sources without copy-pasted tests.

The source-agnostic lifecycle assertions that previously lived in
standalone supernode tests (a valid proposal anchors; an invalid proposal
is beaten, its credit claimed, and never advances the anchor state) now
ride on the supernode subtests of the valid/invalid scenarios, so they
run exactly once instead of duplicating the decision tests. Removes the
redundant standalone supernode valid/invalid tests and the op-node
claims-credit test (credit claiming is source-agnostic and now covered by
the invalid supernode subtest). Addresses review feedback.
